Building Multi Agentic AI RAG With Vector Database/requirements.txt
phidata
duckduckgo-search
fastapi
uvicorn 
groq
openai
sqlalchemy
pgvector
psycopg[binary]
pypdf
streamlit